mm: COW: skip the page lock in the COW copy path
The only guarantee required to be allowed to copy is that mapcount is
not equal 1. That is enough not to risk overestimating mapcount.

This should retain the performance improvement reported below for the
COW copy path, but without introducing any regression to the page
reuse accuracy.

https://lkml.kernel.org/r/20200914024321.GG26874@shao2-debian

Signed-off-by: Andrea Arcangeli <aarcange@redhat.com>
(cherry picked from commit a8c8a8210d6f6a3bbaec1185a1d73f756b17116c)
Signed-off-by: Andrea Arcangeli <andrea@cpushare.com>
Change-Id: I654085cc1a3212eec603f5a1224cee0fdd1b45c6